+++ source/extensions/omni.isaac.ros2_bridge/python/nodes/OgnROS2CameraHelper.py
@@ -81,11 +81,18 @@ class OgnROS2CameraHelper:
                         return False
                 db.per_instance_state.resetSimulationTimeOnStop = db.inputs.resetSimulationTimeOnStop
                 writer = None
+
+                time_type = ""
+                if db.inputs.useSystemTime:
+                    time_type = "SystemTime"
+                    if db.inputs.resetSimulationTimeOnStop:
+                        carb.log_warn("System timestamp is being used. Ignoring resetSimulationTimeOnStop input")
+
                 try:
                     if sensor_type == "rgb":
 
                         rv = omni.syntheticdata.SyntheticData.convert_sensor_type_to_rendervar(sd.SensorType.Rgb.name)
-                        writer = rep.writers.get(rv + "ROS2PublishImage")
+                        writer = rep.writers.get(rv + f"ROS2{time_type}PublishImage")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -98,7 +105,7 @@ class OgnROS2CameraHelper:
                         rv = omni.syntheticdata.SyntheticData.convert_sensor_type_to_rendervar(
                             sd.SensorType.DistanceToImagePlane.name
                         )
-                        writer = rep.writers.get(rv + "ROS2PublishImage")
+                        writer = rep.writers.get(rv + f"ROS2{time_type}PublishImage")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -113,7 +120,7 @@ class OgnROS2CameraHelper:
                             sd.SensorType.DistanceToImagePlane.name
                         )
 
-                        writer = rep.writers.get(rv + "ROS2PublishPointCloud")
+                        writer = rep.writers.get(rv + f"ROS2{time_type}PublishPointCloud")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -124,7 +131,7 @@ class OgnROS2CameraHelper:
 
                     elif sensor_type == "instance_segmentation":
 
-                        writer = rep.writers.get("ROS2PublishInstanceSegmentation")
+                        writer = rep.writers.get(f"ROS2{time_type}PublishInstanceSegmentation")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -134,7 +141,7 @@ class OgnROS2CameraHelper:
                         )
 
                     elif sensor_type == "semantic_segmentation":
-                        writer = rep.writers.get("ROS2PublishSemanticSegmentation")
+                        writer = rep.writers.get(f"ROS2{time_type}PublishSemanticSegmentation")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -144,7 +151,7 @@ class OgnROS2CameraHelper:
                         )
 
                     elif sensor_type == "bbox_2d_tight":
-                        writer = rep.writers.get("ROS2PublishBoundingBox2DTight")
+                        writer = rep.writers.get(f"ROS2{time_type}PublishBoundingBox2DTight")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -154,7 +161,7 @@ class OgnROS2CameraHelper:
                         )
 
                     elif sensor_type == "bbox_2d_loose":
-                        writer = rep.writers.get("ROS2PublishBoundingBox2DLoose")
+                        writer = rep.writers.get(f"ROS2{time_type}PublishBoundingBox2DLoose")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -164,7 +171,7 @@ class OgnROS2CameraHelper:
                         )
 
                     elif sensor_type == "bbox_3d":
-                        writer = rep.writers.get("ROS2PublishBoundingBox3D")
+                        writer = rep.writers.get(f"ROS2{time_type}PublishBoundingBox3D")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -174,7 +181,7 @@ class OgnROS2CameraHelper:
                         )
 
                     elif sensor_type == "camera_info":
-                        writer = rep.writers.get("ROS2PublishCameraInfo")
+                        writer = rep.writers.get(f"ROS2{time_type}PublishCameraInfo")
                         writer.initialize(
                             frameId=db.inputs.frameId,
                             nodeNamespace=db.inputs.nodeNamespace,
@@ -199,7 +206,9 @@ class OgnROS2CameraHelper:
                     }
                     if sensor_type in type_dict:
                         if db.inputs.enableSemanticLabels:
-                            semantic_writer = rep.writers.get(type_dict[sensor_type] + "ROS2PublishSemanticLabels")
+                            semantic_writer = rep.writers.get(
+                                type_dict[sensor_type] + f"ROS2{time_type}PublishSemanticLabels"
+                            )
                             semantic_writer.initialize(
                                 nodeNamespace=db.inputs.nodeNamespace,
                                 queueSize=db.inputs.queueSize,