+++ source/extensions/omni.isaac.ros2_bridge/python/nodes/OgnROS2RtxLidarHelper.py
@@ -82,17 +82,24 @@ class OgnROS2RtxLidarHelper:
                 sensor_type = db.inputs.type
                 db.per_instance_state.resetSimulationTimeOnStop = db.inputs.resetSimulationTimeOnStop
                 writer = None
+
+                time_type = ""
+                if db.inputs.useSystemTime:
+                    time_type = "SystemTime"
+                    if db.inputs.resetSimulationTimeOnStop:
+                        carb.log_warn("System timestamp is being used. Ignoring resetSimulationTimeOnStop input")
+
                 try:
                     if sensor_type == "laser_scan":
                         if db.inputs.fullScan:
                             carb.log_warn("fullScan does not have an effect with the laser_scan setting")
-                        writer = rep.writers.get("RtxLidar" + "ROS2PublishLaserScan")
+                        writer = rep.writers.get("RtxLidar" + f"ROS2{time_type}PublishLaserScan")
 
                     elif sensor_type == "point_cloud":
                         if db.inputs.fullScan:
-                            writer = rep.writers.get("RtxLidar" + "ROS2PublishPointCloudBuffer")
+                            writer = rep.writers.get("RtxLidar" + f"ROS2{time_type}PublishPointCloudBuffer")
                         else:
-                            writer = rep.writers.get("RtxLidar" + "ROS2PublishPointCloud")
+                            writer = rep.writers.get("RtxLidar" + f"ROS2{time_type}PublishPointCloud")
 
                     else:
                         carb.log_error("type is not supported")